+++ src/router/auth.routes.js
@@ -1,0 +1,10 @@
+export const authRoutes = [
+    {
+        path: '/login',
+        name: 'LoginPage',
+        component: () => import('@/views/login/LoginPage.vue'),
+        meta: {
+            hideSidebar: true
+        }
+    }
+];
