+++ src/router/middleware/auth.guard.js
@@ -1,0 +1,15 @@
+import { useAuthStore } from '@/stores/authStore';
+
+export const authGuard = (to, from, next) => {
+    const authStore = useAuthStore();
+    const requiresAuth = to.matched.some(record => record.meta.requiresAuth);
+    const isAuthenticated = authStore.isAuthenticated;
+
+    if (requiresAuth && !isAuthenticated) {
+        next('/login');
+    } else if (to.path === '/login' && isAuthenticated) {
+        next('/');
+    } else {
+        next();
+    }
+};
